pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "google-auth>=2.43.0",
    "google-cloud-discoveryengine>=0.15.0",
    "google-cloud-documentai>=3.7.0",
    "google-cloud-storage>=3.6.0",
    "google-genai>=1.52.0",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "trafilatura>=2.0.0",
]
